requirements.txt
pydantic==2.11.9

langchain==1.1.3
langchain-core==1.1.3
langchain-community==0.4.1
langchain-groq==1.1.0

crewai==1.7.0
crewai-tools>=0.4.0

litellm>=1.45.0

chromadb>=0.5.5
sentence-transformers>=2.7.0
duckduckgo-search>=6.1.0
python-dotenv>=1.0.1
streamlit>=1.40.0
